requirements.txt
# FastAPI Server
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# OCR Engines (Dual: EasyOCR + PaddleOCR, CPU-only)
easyocr==1.7.2
paddleocr==2.10.0
paddlepaddle==3.0.0

# Image Processing
Pillow==11.2.1
pillow-heif==1.3.0
opencv-python-headless==4.11.0.86
numpy==2.2.5

# Database
sqlalchemy==2.0.41
alembic==1.15.2
aiosqlite==0.21.0

# Utilities
pydantic==2.11.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
httpx==0.28.1
